Ensure period is set within the hardware's limitations.
It is possible for an app to send out a notification that has a period of less than 2 seconds or greater than 12 seconds. When that happens, the period is not set at all and causes strange things to happen.

This commit simply ensures that period is always between 2 and 12. Also it fixes a tiny code formatting issue.

Change-Id: I638f7707cb37e1d3008e768638283b463192f70e